migrate clock
