--- src/lib/loading-screen.ts
@@ -1,14 +1,0 @@
-import { ILoadingScreen } from '@babylonjs/core'
-
-export class LoadingScreen implements ILoadingScreen {
-  public loadingUIBackgroundColor: string = '#fff'
-  public loadingUIText: string = ''
-
-  public displayLoadingUI() {
-    return null
-  }
-
-  public hideLoadingUI() {
-    return null
-  }
-}
